DOCSP-3936 Push docs to docs.mongodb.com
This updates the destination to a stitch-sdks subdirectory on the
existing bucket so that docs.mongodb.com/stitch-sdks can find it.
(Not sure, this was specifically required by techops.)

Removes the /docs/ part which is arguably redundant now that it is
hosted on docs.mongodb.com.

Also updates the READMEs for the new docs URL. Please publish
the docs and check that the new URL works.